Refactor out duplicated example rendering
