+++ scripts/typedoc/shopify-dev-renderer/components/utilities/index.ts
@@ -1,5 +1,3 @@
-export {findExamplesForComponent, renderExamplesForComponent} from './examples';
-
 export {
   renderSandboxComponentExamples,
   compileComponentExamples,
