--- scripts/typedoc/shopify-dev-renderer/components/utilities/types.ts
@@ -1,5 +1,0 @@
-export interface Example {
-  filename: string;
-  extension: string;
-  content: string;
-}
